ai_agent/use_cases/task_automation.py
```python
from ai_agent.local_computer_control import executeCommand
from ai_agent.task_automation import scheduleTask
from datetime import datetime

class TaskAutomationUseCase:
    def __init__(self):
        self.taskQueue = []

    def automate_file_organization(self, source_folder, destination_folder):
        command = f'move {source_folder}/* {destination_folder}'
        executeCommand(command)
        self.taskQueue.append({'task': 'file organization', 'status': 'completed', 'timestamp': datetime.now()})

    def automate_email_management(self, email_account, action, condition):
        task = {
            'email_account': email_account,
            'action': action,
            'condition': condition
        }
        scheduleTask('email management', task)
        self.taskQueue.append({'task': 'email management', 'status': 'scheduled', 'timestamp': datetime.now()})

    def automate_data_processing(self, data_file, process):
        command = f'{process} {data_file}'
        executeCommand(command)
        self.taskQueue.append({'task': 'data processing', 'status': 'completed', 'timestamp': datetime.now()})
```
